requirement.txt
streamlit
pandas
numpy
PIL 
seaborn
matplotlib.pyplot